Refractored project to more MVC style
