+++ .gitignore
@@ -1,3 +1,3 @@
-/venv/
-app/__pycache__
-/data/excuses_ranks.db
+venv
+__pycache__
+excuses_ranks.db
