--- app/controller.py
@@ -1,12 +1,0 @@
-from fastapi import Request
-import model
-import view
-
-async def root_controller(request: Request):
-    return view.render_home_page(request)
-
-async def change_excuse_rank(text: str, deltaRank: int):
-    model.change_rank(text, deltaRank)
-
-async def get_excuses(numberOfExcuses: int, needAll: bool):
-    return model.get_excuses(numberOfExcuses, needAll)
