+++ src/controllers/controller.py
@@ -1,0 +1,21 @@
+from fastapi import Request, APIRouter
+from fastapi.responses import HTMLResponse, JSONResponse
+import src.models.model as model
+import src.views.view as view
+
+router = APIRouter()
+
+#Root page
+@router.get("/", response_class=HTMLResponse)
+async def root(request: Request):
+    return view.render_home_page(request)
+
+#Change excuses rank
+@router.post("/change-excuse-rank",)
+async def change_excuse_rank(text: str, deltaRank = 1):
+    model.change_rank(text, deltaRank)
+
+#Get some excuses in sorted order (by rank)
+@router.get("/get-excuses", response_class=JSONResponse)
+async def get_excuses(numberOfExcuses = 0, needAll: bool = False):
+    return model.get_excuses(numberOfExcuses, needAll)
